+++ src/main/java/com/zephyr/croj/bootstrap/AdminBootstrapRequest.java
@@ -1,0 +1,9 @@
+package com.zephyr.croj.bootstrap;
+
+public record AdminBootstrapRequest(String username, String email, String password) {
+
+    @Override
+    public String toString() {
+        return "AdminBootstrapRequest[username=<redacted>, email=<redacted>, password=<redacted>]";
+    }
+}